attenuation

+++ scripts/utils.py
@@ -20,7 +20,7 @@ def activation_opt_flow(x):
     return tf.stack((x0,x1,x2),axis=-1)
 
 def sensory_attenuation(predicted_opt_flow, next_image, background_image):
-    unnorm_pred = np.zeros(next_image.shape, dtype=b.dtype) + (1.0 - predicted_opt_flow)*255
+    unnorm_pred = np.zeros(next_image.shape, dtype=next_image.dtype) + (1.0 - predicted_opt_flow)*255
     print('unnomr shape ', unnorm_pred.shape)
     print('next_image shape ', next_image.shape)
     print('background_image shape ', background_image.shape)
